requirements.txt
aiodns>=3.1.1
aiofiles>=23.2.1
alive_progress>=3.1.4
art>=6.1
async_dns>=2.0.0
colorama>=0.4.4
dnspython>=2.5.0
Requests>=2.31.0
urllib3>=1.26.18
